Refactor plugin EP support (#25541)
### Description
<!-- Describe your changes. -->
Refactor to split out classes and make things easier to find. 

### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->
Cleanup
